Added insert tag extension added in tags 1.7.5
Extension of the insert tags tags_news, tags_event, tags_article,
tags_content by additional parameters. The insert tags can be used now
with the following parameters:
{{tags_article::ID::MAX::RELEVANCE::TARGET}}, where ID is the ID of the
news entry, the event, the article, or the content element. Optionally
you can use MAX = maximum number of tags, RELEVANCE = 1 is sorting the
tags by relevance, otherwise in alphabetical order, TARGET = ID or
Alias of a contao page to jump to. In case of a redirection the tags
will be shown with an URL. Example:
{{tags_article::5::0::1::articlelist}}: This shows all tags of the
article with the ID 5, sorted by relevance. The tags are linked and
jumpt to the page with the alias articlelist.